bot_🤖/requirements.txt
colorthief==0.2.1
Pillow==10.0.0
Requests==2.31.0
tweepy==4.14.0
